# docs/openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHealth","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/clients":{"post":{"operationId":"ClientsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["clients"]},"get":{"operationId":"ClientsController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["clients"]}},"/clients/{id}":{"get":{"operationId":"ClientsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["clients"]}},"/clients/{clientId}/documents":{"post":{"operationId":"DocumentsController_create","parameters":[{"name":"clientId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDocumentDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]},"get":{"operationId":"DocumentsController_findByClient","parameters":[{"name":"clientId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}},"/documents/{id}":{"get":{"operationId":"DocumentsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}},"/documents/{id}/file":{"get":{"operationId":"DocumentsController_downloadFile","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}},"/documents/{id}/upload":{"post":{"operationId":"DocumentsController_upload","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}},"/documents/{id}/start-review":{"post":{"operationId":"DocumentsController_startReview","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}},"/documents/{id}/review":{"post":{"operationId":"DocumentsController_review","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewDocumentDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}},"/audit-log":{"get":{"operationId":"DocumentsController_firmAuditLog","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}},"/documents/{id}/audit-log":{"get":{"operationId":"DocumentsController_auditLog","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}},"/clients/{clientId}/audit-log":{"get":{"operationId":"DocumentsController_clientAuditLog","parameters":[{"name":"clientId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["documents"]}}},"info":{"title":"OBLIQ-in Audit Workflow API","description":"Mini Audit Document Review System — backend prototype","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"rohit@abc-co.test"},"password":{"type":"string","example":"password123"}},"required":["email","password"]},"CreateClientDto":{"type":"object","properties":{"name":{"type":"string","example":"ABC Traders Pvt. Ltd."}},"required":["name"]},"CreateDocumentDto":{"type":"object","properties":{"name":{"type":"string","example":"Bank Statement","description":"The required document type/name for this client."}},"required":["name"]},"ReviewDocumentDto":{"type":"object","properties":{"action":{"type":"string","enum":["APPROVE","REQUEST_CORRECTION"]},"comment":{"type":"string","example":"Page 3 is missing. Please upload the complete bank statement.","description":"Required when action is REQUEST_CORRECTION."}},"required":["action"]}}}}